# .well-known/openapi.yaml
components:
  schemas:
    EvaluateMovesRequest:
      properties:
        pgn:
          title: Pgn
          type: string
      required:
      - pgn
      title: EvaluateMovesRequest
      type: object
    Evaluation:
      properties:
        pgn_with_score_change:
          title: Pgn With Score Change
          type: string
      required:
      - pgn_with_score_change
      title: Evaluation
      type: object
    GenerateGifGameRequest:
      properties:
        pgn_data:
          title: Pgn Data
          type: string
        reverse:
          default: false
          title: Reverse
          type: boolean
      required:
      - pgn_data
      title: GenerateGifGameRequest
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    SearchRequest:
      properties:
        color:
          title: Color
          type: string
        name:
          title: Name
          type: string
        username_chess_com:
          title: Username Chess Com
          type: string
      title: SearchRequest
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: My Application
  version: 1.0.0
openapi: 3.0.2
paths:
  /evaluate_moves:
    post:
      operationId: evaluate_moves_evaluate_moves_post
      parameters:
      - in: query
        name: description
        required: false
        schema:
          default: Evaluate score change for each moves. Only show change if > 15
          title: Description
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateMovesRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evaluation'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Evaluate Moves
  /games:
    post:
      operationId: search_games_games_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Search Games
  /generate-gif-game/:
    post:
      description: Generate a gif game from a PGN file. Use reverse if player is black.
        Embed the gif in markdown with ![]({url})
      operationId: generate_gif_game_generate_gif_game__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateGifGameRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Generate Gif Game
  /generate-openapi-yaml:
    get:
      operationId: generate_openapi_yaml_generate_openapi_yaml_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Generate Openapi Yaml